#NewFeatures
This update is for established a new maintenance code, using the
programming model MVC. I create a new class defined like as controller,
that control all instance of the game. With this the ViewController for
the GUI and the Controller for the game is in code divided. Later i
update to make the code work